requirements.txt
fastapi
uvicorn
python-multipart
pypdf
mcp
groq
python-dotenv
